+++ app/Providers/AppServiceProvider.php
@@ -4,12 +4,15 @@ declare(strict_types=1);
 
 namespace App\Providers;
 
-use Illuminate\Support\ServiceProvider;
+use Illuminate\Cache\RateLimiting\Limit;
 use Illuminate\Database\Events\MigrationsEnded;
 use Illuminate\Database\Events\MigrationsStarted;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Event;
+use Illuminate\Support\Facades\RateLimiter;
 use Illuminate\Support\Facades\URL;
+use Illuminate\Support\ServiceProvider;
 
 final class AppServiceProvider extends ServiceProvider
 {
@@ -31,6 +34,10 @@ final class AppServiceProvider extends ServiceProvider
         if (config('app.force_https')) {
             URL::forceScheme('https');
         }
+
+        RateLimiter::for('api', function (Request $request) {
+            return Limit::perMinute(60)->by($request->user()?->id ?: $request->ip());
+        });
     }
 
     /**
@@ -41,11 +48,11 @@ final class AppServiceProvider extends ServiceProvider
      */
     private function allowDisabledPk(): void
     {
-        Event::listen(MigrationsStarted::class, function () {
+        Event::listen(MigrationsStarted::class, function (): void {
             DB::statement('SET SESSION sql_require_primary_key=0');
         });
 
-        Event::listen(MigrationsEnded::class, function () {
+        Event::listen(MigrationsEnded::class, function (): void {
             DB::statement('SET SESSION sql_require_primary_key=1');
         });
     }